+++ packages/block-editor/src/components/provider/use-media-upload-settings.js
@@ -20,6 +20,8 @@ function useMediaUploadSettings( settings = {} ) {
 			bigImageSizeThreshold: settings.bigImageSizeThreshold,
 			imageStripMeta: settings.imageStripMeta,
 			imageMaxBitDepth: settings.imageMaxBitDepth,
+			generateAnimatedImageSubsizes:
+				settings.generateAnimatedImageSubsizes,
 		} ),
 		[ settings ]
 	);